+++ lib/modules/initial/pages/main/initial.page.dart
@@ -12,8 +12,8 @@ class InitialPageContianer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return WillPopScope(
-        onWillPop: () async => false,
+    return PopScope(
+        onPopInvoked: (_) => false,
         child: MultiProvider(
           providers: [
             ChangeNotifierProvider<ManifestDownloaderNotifier>(