+++ windows/runner/main.cpp
@@ -4,8 +4,7 @@
 
 #include "flutter_window.h"
 #include "utils.h"
-
-#include <uni_links_desktop/uni_links_desktop_plugin.h>
+#include "app_links/app_links_plugin_c_api.h"
 
 int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
                       _In_ wchar_t *command_line, _In_ int show_command)
@@ -13,7 +12,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE instance, _In_opt_ HINSTANCE prev,
   HWND hwnd = ::FindWindow(L"FLUTTER_RUNNER_WIN32_WINDOW", L"littlelight");
   if (hwnd != NULL)
   {
-    DispatchToUniLinksDesktop(hwnd);
+    SendAppLink(hwnd);
     ::ShowWindow(hwnd, SW_NORMAL);
     ::SetForegroundWindow(hwnd);
     return EXIT_FAILURE;